Modularize Transforms (#264)
