+++ weld/transforms/loop_fusion.rs
@@ -1,253 +1,22 @@
-//! Common transformations on expressions.
+//! Transforms which fuse loops to reduce memory movement and prevent unncessary
+//! traversals of data.
 
-use super::ast::*;
-use super::ast::ExprKind::*;
-use super::ast::Type::*;
-use super::ast::BuilderKind::*;
-use super::ast::LiteralKind::*;
-use super::error::*;
-use super::annotations::*;
-use super::exprs;
+use ast::*;
+use ast::ExprKind::*;
+use ast::Type::*;
+use ast::BuilderKind::*;
+use ast::LiteralKind::*;
+use error::*;
+use annotations::*;
+use exprs;
 
-use super::util::SymbolGenerator;
+use super::inliner::inline_apply;
 
-use std::collections::HashMap;
-use std::collections::hash_map::Entry;
+use util::SymbolGenerator;
 
-/// Inlines Zip expressions as collections of iters. Using Zips outside of a For loop is currently
-/// unsupported behavior. This transform handles the simple case of converting Zips in macros
-/// such as map and filter into Iters in For loops.
-///
-/// TODO(shoumik): Perhaps Zip should just be a macro? Then macros need to be ordered.
-pub fn inline_zips(expr: &mut Expr<Type>) {
-    expr.transform(&mut |ref mut e| {
-        if let For {
-                   ref mut iters,
-                   ref builder,
-                   ref func,
-               } = e.kind {
-            if iters.len() == 1 {
-                let ref first_iter = iters[0];
-                if let Zip { ref vectors } = first_iter.data.kind {
-                    let new_iters = vectors
-                        .iter()
-                        .map(|v| {
-                                 Iter {
-                                     data: Box::new(v.clone()),
-                                     start: None,
-                                     end: None,
-                                     stride: None,
-                                     kind: first_iter.kind.clone(),
-                                 }
-                             })
-                        .collect::<Vec<_>>();
-                    return Some(Expr {
-                                    ty: e.ty.clone(),
-                                    kind: For {
-                                        iters: new_iters,
-                                        builder: builder.clone(),
-                                        func: func.clone(),
-                                    },
-                                    annotations: Annotations::new(),
-                                });
-                }
-            }
-        }
-        None
-    });
-}
-
-/// A stack which keeps track of unique variables names and scoping information for symbols.
-struct SymbolStack {
-    // The symbol stack.
-    stack: HashMap<Symbol, Vec<i32>>,
-    // The next unique ID to assign to this name.
-    next_unique_symbol: HashMap<String, i32>,
-}
-
-impl SymbolStack {
-    fn new() -> SymbolStack {
-        SymbolStack {
-            stack: HashMap::new(),
-            next_unique_symbol: HashMap::new(),
-        }
-    }
-
-    /// Returns the symbol in the current scope for the given symbol, or an error if the symbol is
-    /// undefined.
-    fn symbol(&mut self, sym: Symbol) -> WeldResult<Symbol> {
-        match self.stack.entry(sym.clone()) {
-            Entry::Occupied(ref ent) => {
-                let name = ent.key().name.as_str();
-                let id = ent.get()
-                    .last()
-                    .map(|v| *v)
-                    .ok_or(WeldError::new(format!("Symbol {} is out of scope", &sym)))?;
-                Ok(Symbol::new(name, id))
-            }
-            _ => weld_err!("Undefined symbol {}", sym),
-        }
-    }
-
-
-    /// Push a new symbol onto the stack, assigning it a unique name. This enters a new scope for
-    /// the name. The symbol can be retrieved with `symbol()`.
-    fn push_symbol(&mut self, sym: Symbol) {
-        let mut stack_entry = self.stack.entry(sym.clone()).or_insert(Vec::new());
-        let mut next_entry = self.next_unique_symbol.entry(sym.name).or_insert(-1);
-        *next_entry = if sym.id > *next_entry {
-            sym.id
-        } else {
-           *next_entry + 1 
-        };
-        stack_entry.push(*next_entry);
-    }
-
-    /// Pop a symbol from the stack.
-    fn pop_symbol(&mut self, sym: Symbol) -> WeldResult<()> {
-        match self.stack.entry(sym.clone()) {
-            Entry::Occupied(mut ent) => {
-                ent.get_mut().pop();
-                Ok(())
-            },
-            _ => weld_err!("Attempting to pop undefined symbol {}", sym)
-        }
-    }
-}
-
-/// Modifies symbol names so each symbol is unique in the AST.  Returns an error if an undeclared
-/// symbol appears in the program.
-pub fn uniquify<T: TypeBounds>(expr: &mut Expr<T>) -> WeldResult<()> {
-    uniquify_helper(expr, &mut SymbolStack::new())
-}
-
-/// The main helper function for uniquify, which uses `SymbolStack` to track scope and assign
-/// unique names to each symbol. The prerequisite is that each symbol has `id = 0`.
-fn uniquify_helper<T: TypeBounds>(expr: &mut Expr<T>, symbol_stack: &mut SymbolStack) -> WeldResult<()> {
-    match expr.kind {
-        // First, handle expressions which define *new* symbols - Let and Lambda
-        Lambda {ref mut params, ref mut body} => {
-            // Update the parameter of the lambda with new names.
-            let original_params = params.clone();
-            for param in params.iter_mut() {
-                let ref mut sym = param.name;
-                symbol_stack.push_symbol(sym.clone());
-                *sym = symbol_stack.symbol(sym.clone())?;
-            }
-
-            // Then, uniquify the lambda using the newly pushed symbols.
-            uniquify_helper(body, symbol_stack)?;
-            
-            // Finally, pop off the symbol names since they are out of scope now.
-            for param in original_params {
-                symbol_stack.pop_symbol(param.name)?;
-            }
-        }
-        Let {ref mut name, ref mut value, ref mut body} => {
-            // First, uniquify the value *without* the updated stack, since the Let hasn't defined
-            // the symbol yet.
-            uniquify_helper(value, symbol_stack)?;
-
-            // Now, push the Let's symbol name.
-            symbol_stack.push_symbol(name.clone());
-            let original_name = name.clone();
-            *name = symbol_stack.symbol(name.clone())?;
-
-            // uniquify the body with the new scope.
-            uniquify_helper(body, symbol_stack)?;
-
-            // Pop off the scope.
-            symbol_stack.pop_symbol(original_name)?;
-        }
-        // Now handle identifiers, which are changed to reflect new symbols.
-        Ident(ref mut sym) => {
-            *sym = symbol_stack.symbol(sym.clone())?;
-        }
-        // For all other expressions, call uniquify_helper on the children.
-        _ => {
-            for child in expr.children_mut() {
-                uniquify_helper(child, symbol_stack)?;
-            }
-        }
-    }
-    Ok(())
-}
-
-/// Inlines Apply nodes whose argument is a Lambda expression. These often arise during macro
-/// expansion but it's simpler to inline them before doing type inference.
-/// Unlike many of the other transformations, we make this one independent of types so that
-/// we can apply it before type inference.
-///
-/// Caveats:
-/// - Functions that reuse a parameter twice have its expansion appear twice, instead of assigning
-///   it to a temporary as would happen with function application.
-/// - Does not complete inlining if some of the functions take functions as arguments (in that
-///   case, the expressions after inlining may lead to more inlining).
-pub fn inline_apply<T: TypeBounds>(expr: &mut Expr<T>) {
-    expr.transform(&mut |ref mut expr| {
-        if let Apply {
-                   ref func,
-                   params: ref args,
-               } = expr.kind {
-            if let Lambda {
-                       ref params,
-                       ref body,
-                   } = func.kind {
-                let mut new = *body.clone();
-                for (param, arg) in params.iter().zip(args) {
-                    new.substitute(&param.name, &arg);
-                }
-                return Some(new);
-            }
-        }
-        None
-    });
-}
-
-/// Inlines Let calls if the symbol defined by the Let statement is used
-/// less than one time.
-pub fn inline_let(expr: &mut Expr<Type>) {
-    if let Ok(_) = uniquify(expr) {
-        expr.transform(&mut |ref mut expr| {
-            if let Let {
-                ref mut name,
-                ref mut value,
-                ref mut body,
-            } = expr.kind {
-                if symbol_usage_count(name, body) <= 1 {
-                    body.transform(&mut |ref mut expr| {
-                        if let Ident(ref symbol) = expr.kind {
-                            if symbol == name {
-                                return Some(*value.clone());
-                            }
-                        }
-                        return None;
-                    });
-                    return Some(*body.clone());
-                }
-            }
-            return None;
-        });
-    }
-}
-
-fn symbol_usage_count(sym: &Symbol, expr: &Expr<Type>) -> u32 {
-    let mut usage_count = 0;
-    expr.traverse(&mut |ref e| {
-        if let For { ref func, .. } = e.kind {
-            // The other child expressions of the For will be counted by traverse.
-            if symbol_usage_count(sym, func) >= 1 {
-                usage_count += 3;
-            }
-        } else if let Ident(ref symbol) = e.kind {
-            if sym == symbol {
-                usage_count += 1;
-            }
-        }
-    });
-
-    usage_count
-}
+/// Fuses loops where one for loop takes another as it's input, which prevents intermediate results
+/// from being materialized.
+pub struct VerticalLoopFusion;
 
 /// Fuses for loops over the same vector in a zip into a single for loop which produces a vector of
 /// structs directly.
@@ -563,101 +332,4 @@ fn match_types(root_ty: &Type, expr: &mut Expr<Type>) {
         }
         _ => {}
     };
-}
-
-/// Simplifies GetField(MakeStruct(*)) expressions, which can occur during loop fusion when some
-/// of the loops are zipping together multiple column vectors.
-pub fn simplify_get_field<T: TypeBounds>(expr: &mut Expr<T>) {
-    expr.transform(&mut |ref mut expr| {
-        if let GetField { ref expr, index } = expr.kind {
-            if let MakeStruct { ref elems } = expr.kind {
-                return Some(elems[index as usize].clone());
-            }
-        }
-        None
-    });
-}
-
-/// Infers the size of an `Appender` in a `For` loop.
-pub fn infer_size(expr: &mut Expr<Type>) {
-    expr.transform_and_continue_res(&mut |ref mut expr| {
-        if let For { ref iters, ref builder, ref func } = expr.kind {
-            // This constraint prevents copying expensive iters.
-            if let Ident(_) = iters[0].data.kind {
-                if let NewBuilder(None) = builder.kind {
-                    if let Builder(Appender(ref ek), _) = builder.ty {
-                        if let Lambda {ref params, ref body } = func.kind {
-                            let ref builder_symbol = params[0].name.clone();
-                            if simple_merge(builder_symbol, body) {
-                                // Compute the inferred length based on the iter.
-                                let length = if let Some(ref start) = iters[0].start {
-                                    let e = exprs::binop_expr(BinOpKind::Subtract,
-                                                              *iters[0].end.as_ref().unwrap().clone(),
-                                                              *start.clone())?;
-                                    exprs::binop_expr(BinOpKind::Divide, e, *iters[0].stride.as_ref().unwrap().clone())?
-                                } else {
-                                    exprs::length_expr(*iters[0].data.clone())?
-                                };
-
-                                let new_loop = exprs::for_expr(
-                                    iters.clone(),
-                                    exprs::newbuilder_expr(Appender(ek.clone()), Some(length))?,
-                                    func.as_ref().clone(),
-                                    false)?;
-                                return Ok((Some(new_loop), false));
-                            }
-                        }
-                    }
-                }
-            }
-        }
-        Ok((None, true))
-    })
-}
-
-/// Checks that `expr` performs only one `Merge` per control path - this guarantees
-/// that the loop containing `expr`'s size can be inferred.
-fn simple_merge(sym: &Symbol, expr: &Expr<Type>) -> bool {
-    match expr.kind {
-        Merge { ref builder, ref value } => {
-            if let Ident(ref s) = builder.kind {
-                if s == sym {
-                    return !value.contains_symbol(sym);
-                }
-            }
-            return false;
-        }
-        If { ref cond, ref on_true, ref on_false } => {
-            !cond.contains_symbol(sym) && simple_merge(sym, on_true) &&
-                simple_merge(sym, on_false)
-        }
-        _ => false,
-    }
-}
-
-/// Forces parallelization of for loops inside an `Iterate`. In the current LLVM code generator,
-/// this prevents for loop continuations from making recursive calls to the beginning of the
-/// iteration, which can cause stack exhaustion.
-///
-/// FIXME we need a long term solution to this!
-pub fn force_iterate_parallel_fors(expr: &mut Expr<Type>) {
-    expr.transform_and_continue(&mut |ref mut e| {
-        match e.kind {
-             Iterate { .. } => {
-                e.transform_and_continue(&mut |ref mut e| {
-                   match e.kind {
-                        For { .. } => {
-                            e.annotations.set_always_use_runtime(true);
-                            (None, true)
-                        }
-                        _ => (None, true)
-                   }
-                });
-                (None, false)
-            }
-            _ => {
-                (None, true)
-            }
-        }
-    });
 }
