+++ weld/transforms/vectorizer.rs
@@ -6,17 +6,161 @@
 
 use std::collections::HashSet;
 
-use super::ast::*;
-use super::ast::ExprKind::*;
-use super::ast::Type::*;
-use super::error::*;
-use super::exprs;
-use super::util::SymbolGenerator;
+use ast::*;
+use ast::ExprKind::*;
+use ast::Type::*;
+use error::*;
+use exprs;
+use util::SymbolGenerator;
 
 #[cfg(test)]
-use super::parser::*;
+use parser::*;
 #[cfg(test)]
-use super::type_inference::*;
+use type_inference::*;
+
+/// Vectorize an expression.
+pub fn vectorize(expr: &mut Expr<Type>) {
+    let mut vectorized = false;
+    // Used to create the identifiers which refer to the data items. These identifiers are
+    // used to pull out the iter into a let statement. This lets us repeat the iter via an
+    // identifier in the vectorized loop later. Declaring this before any transformations so
+    // there is no clash of variable names.
+    let mut sym_gen = SymbolGenerator::from_expression(expr);
+
+    expr.transform_and_continue_res(&mut |ref mut expr| {
+        //  The Res is a stricter-than-necessary check, but prevents us from having to check nested
+        //  loops for now.
+        if let Res { builder: ref for_loop } = expr.kind {
+            let ref broadcast_idens = vectorizable(for_loop)?;
+            if let For { ref iters, builder: ref init_builder, ref func } = for_loop.kind {
+                if let NewBuilder(_) = init_builder.kind {
+                    if let Lambda { ref params, ref body } = func.kind {
+                        // This is the vectorized body.
+                        let mut vectorized_body = body.clone();
+                        vectorized_body.transform_and_continue(&mut |ref mut e| {
+                                                                        let cont = vectorize_expr(e, broadcast_idens).unwrap();
+                                                                        (None, cont)
+                                                                    });
+
+                        let mut vectorized_params = params.clone();
+                        vectorized_params[2].ty = vectorized_params[2].ty.simd_type()?;
+
+                        let vec_func = exprs::lambda_expr(vectorized_params, *vectorized_body)?;
+
+                        let data_names = iters
+                            .iter()
+                            .map(|_| sym_gen.new_symbol("a"))
+                            .collect::<Vec<_>>();
+
+                        // Iterators for the vectorized loop.
+                        let mut vec_iters = vec![];
+                        for (e, n) in iters.iter().zip(&data_names) {
+                            vec_iters.push(Iter {
+                                               data: Box::new(exprs::ident_expr(n.clone(), e.data.ty.clone())?),
+                                               start: e.start.clone(),
+                                               end: e.end.clone(),
+                                               stride: e.stride.clone(),
+                                               kind: IterKind::SimdIter,
+                                           });
+                        }
+
+                        // Iterators for the fringe loop. This is the same set of iterators, but with the
+                        // IteratorKind changed to Fringe.
+                        let fringe_iters = vec_iters
+                            .iter_mut()
+                            .map(|i| {
+                                     let mut i = i.clone();
+                                     i.kind = IterKind::FringeIter;
+                                     i
+                                 })
+                            .collect();
+
+                        let vectorized_loop = exprs::for_expr(vec_iters, *init_builder.clone(), vec_func, true)?;
+                        let scalar_loop = exprs::for_expr(fringe_iters, vectorized_loop, *func.clone(), false)?;
+                        let result = exprs::result_expr(scalar_loop)?;
+
+                        let mut prev_expr = result;
+                        for (iter, name) in iters.iter().zip(data_names).rev() {
+                            prev_expr = exprs::let_expr(name.clone(), *iter.data.clone(), prev_expr)?;
+                        }
+
+                        vectorized = true;
+                        return Ok((Some(prev_expr), false));
+                    }
+                }
+            }
+        }
+        // Check other expressions.
+        Ok((None, true))
+    });
+}
+
+/// Predicate an `If` expression by checking for if(cond, merge(b, e), b) and transforms it to merge(b, select(cond, e,identity)).
+pub fn predicate(e: &mut Expr<Type>) {
+    e.transform_and_continue_res(&mut |ref mut e| {
+        if !(should_be_predicated(e)) {
+            return Ok((None, true));
+        }
+
+        // Predication for a value merged into a merger. This pattern checks for if(cond, merge(b, e), b).
+        if let If { ref cond, ref on_true, ref on_false } = e.kind {
+            if let Merge { ref builder, ref value } = on_true.kind {
+                if let Ident(ref name) = on_false.kind {
+                    if let Ident(ref name2) = builder.kind {
+                        if name == name2 {
+                            if let Builder(ref bk, _) = builder.ty {
+                                // Merge in the identity element if the predicate fails (effectively merging in nothing)
+                                let (ty, op) = match *bk {
+                                    BuilderKind::Merger(ref ty, ref op) => (ty, op),
+                                    BuilderKind::DictMerger(_, ref ty2, ref op) => (ty2, op),
+                                    BuilderKind::VecMerger(ref ty, ref op) => (ty, op),
+                                    _ => {
+                                        return Ok((None, true));
+                                    }
+                                };
+
+                                let identity = get_id_element(ty.as_ref(), op)?;
+                                match identity {
+                                    Some(x) => {
+                                        match *bk {
+                                            BuilderKind::Merger(_, _)
+                                                | BuilderKind::VecMerger(_, _) => {
+                                                /* Change if(cond, merge(b, e), b) => 
+                                                merge(b, select(cond, e, identity). */
+                                                let expr = exprs::merge_expr(*builder.clone(),
+                                                                             exprs::select_expr(
+                                                                                 *cond.clone(),
+                                                                                 *value.clone(), x)?)?;
+                                                return Ok((Some(expr), true));
+                                                
+                                            },
+                                            BuilderKind::DictMerger(_, _, _) => {
+                                                /* For dictmerger, need to match identity element 
+                                                back to the key. */
+                                                let sel_expr = make_select_for_kv(*cond.clone(),
+                                                                                  *value.clone(),
+                                                                                  x)?;
+                                                return Ok((sel_expr, true))
+                                            }
+                                            _ => {
+                                                return Ok((None, true));
+                                            }
+                                        }
+                                    }
+                                    None => {
+                                        return Ok((None, true));
+                                    }
+                                };
+
+                            }
+                        }
+                    }
+                }
+            }
+        }
+        Ok((None, true))
+    });
+}
 
 /// Returns `true` if this is a set of iterators we can vectorize, `false` otherwise.
 ///
@@ -86,7 +230,6 @@ fn vectorize_expr(e: &mut Expr<Type>, broadcast_idens: &HashSet<Symbol>) -> Weld
     }
     Ok(cont)
 }
-
 
 /// Checks basic vectorizability for a loop - this is a strong check which ensure that the only
 /// expressions which appear in a function body are vectorizable expressions (see
@@ -193,11 +336,11 @@ fn vectorizable(for_loop: &Expr<Type>) -> WeldResult<HashSet<Symbol>> {
     return weld_err!("Unsupported pattern");
 }
 
-pub fn should_be_predicated(e: &mut Expr<Type>) -> bool {
+fn should_be_predicated(e: &mut Expr<Type>) -> bool {
     e.annotations.predicate()
 }
 
-pub fn get_id_element(ty: &Type, op: &BinOpKind) -> WeldResult<Option<Expr<Type>>> {
+fn get_id_element(ty: &Type, op: &BinOpKind) -> WeldResult<Option<Expr<Type>>> {
     let ref sk = match *ty {
         Scalar(sk) => sk,
         _ => {
@@ -253,149 +396,7 @@ fn make_select_for_kv(cond:  Expr<Type>,
     return Ok(Some(le))
 }
 
-/// Predicate an `If` expression by checking for if(cond, merge(b, e), b) and transforms it to merge(b, select(cond, e,identity)).
-pub fn predicate(e: &mut Expr<Type>) {
-    e.transform_and_continue_res(&mut |ref mut e| {
-        if !(should_be_predicated(e)) {
-            return Ok((None, true));
-        }
 
-        // Predication for a value merged into a merger. This pattern checks for if(cond, merge(b, e), b).
-        if let If { ref cond, ref on_true, ref on_false } = e.kind {
-            if let Merge { ref builder, ref value } = on_true.kind {
-                if let Ident(ref name) = on_false.kind {
-                    if let Ident(ref name2) = builder.kind {
-                        if name == name2 {
-                            if let Builder(ref bk, _) = builder.ty {
-                                // Merge in the identity element if the predicate fails (effectively merging in nothing)
-                                let (ty, op) = match *bk {
-                                    BuilderKind::Merger(ref ty, ref op) => (ty, op),
-                                    BuilderKind::DictMerger(_, ref ty2, ref op) => (ty2, op),
-                                    BuilderKind::VecMerger(ref ty, ref op) => (ty, op),
-                                    _ => {
-                                        return Ok((None, true));
-                                    }
-                                };
-
-                                let identity = get_id_element(ty.as_ref(), op)?;
-                                match identity {
-                                    Some(x) => {
-                                        match *bk {
-                                            BuilderKind::Merger(_, _)
-                                                | BuilderKind::VecMerger(_, _) => {
-                                                /* Change if(cond, merge(b, e), b) => 
-                                                merge(b, select(cond, e, identity). */
-                                                let expr = exprs::merge_expr(*builder.clone(),
-                                                                             exprs::select_expr(
-                                                                                 *cond.clone(),
-                                                                                 *value.clone(), x)?)?;
-                                                return Ok((Some(expr), true));
-                                                
-                                            },
-                                            BuilderKind::DictMerger(_, _, _) => {
-                                                /* For dictmerger, need to match identity element 
-                                                back to the key. */
-                                                let sel_expr = make_select_for_kv(*cond.clone(),
-                                                                                  *value.clone(),
-                                                                                  x)?;
-                                                return Ok((sel_expr, true))
-                                            }
-                                            _ => {
-                                                return Ok((None, true));
-                                            }
-                                        }
-                                    }
-                                    None => {
-                                        return Ok((None, true));
-                                    }
-                                };
-
-                            }
-                        }
-                    }
-                }
-            }
-        }
-        Ok((None, true))
-    });
-}
-
-/// Vectorize an expression.
-pub fn vectorize(expr: &mut Expr<Type>) {
-    let mut vectorized = false;
-    // Used to create the identifiers which refer to the data items. These identifiers are
-    // used to pull out the iter into a let statement. This lets us repeat the iter via an
-    // identifier in the vectorized loop later. Declaring this before any transformations so
-    // there is no clash of variable names.
-    let mut sym_gen = SymbolGenerator::from_expression(expr);
-
-    expr.transform_and_continue_res(&mut |ref mut expr| {
-        //  The Res is a stricter-than-necessary check, but prevents us from having to check nested
-        //  loops for now.
-        if let Res { builder: ref for_loop } = expr.kind {
-            let ref broadcast_idens = vectorizable(for_loop)?;
-            if let For { ref iters, builder: ref init_builder, ref func } = for_loop.kind {
-                if let NewBuilder(_) = init_builder.kind {
-                    if let Lambda { ref params, ref body } = func.kind {
-                        // This is the vectorized body.
-                        let mut vectorized_body = body.clone();
-                        vectorized_body.transform_and_continue(&mut |ref mut e| {
-                                                                        let cont = vectorize_expr(e, broadcast_idens).unwrap();
-                                                                        (None, cont)
-                                                                    });
-
-                        let mut vectorized_params = params.clone();
-                        vectorized_params[2].ty = vectorized_params[2].ty.simd_type()?;
-
-                        let vec_func = exprs::lambda_expr(vectorized_params, *vectorized_body)?;
-
-                        let data_names = iters
-                            .iter()
-                            .map(|_| sym_gen.new_symbol("a"))
-                            .collect::<Vec<_>>();
-
-                        // Iterators for the vectorized loop.
-                        let mut vec_iters = vec![];
-                        for (e, n) in iters.iter().zip(&data_names) {
-                            vec_iters.push(Iter {
-                                               data: Box::new(exprs::ident_expr(n.clone(), e.data.ty.clone())?),
-                                               start: e.start.clone(),
-                                               end: e.end.clone(),
-                                               stride: e.stride.clone(),
-                                               kind: IterKind::SimdIter,
-                                           });
-                        }
-
-                        // Iterators for the fringe loop. This is the same set of iterators, but with the
-                        // IteratorKind changed to Fringe.
-                        let fringe_iters = vec_iters
-                            .iter_mut()
-                            .map(|i| {
-                                     let mut i = i.clone();
-                                     i.kind = IterKind::FringeIter;
-                                     i
-                                 })
-                            .collect();
-
-                        let vectorized_loop = exprs::for_expr(vec_iters, *init_builder.clone(), vec_func, true)?;
-                        let scalar_loop = exprs::for_expr(fringe_iters, vectorized_loop, *func.clone(), false)?;
-                        let result = exprs::result_expr(scalar_loop)?;
-
-                        let mut prev_expr = result;
-                        for (iter, name) in iters.iter().zip(data_names).rev() {
-                            prev_expr = exprs::let_expr(name.clone(), *iter.data.clone(), prev_expr)?;
-                        }
-
-                        vectorized = true;
-                        return Ok((Some(prev_expr), false));
-                    }
-                }
-            }
-        }
-        // Check other expressions.
-        Ok((None, true))
-    });
-}
 
 /// Parse and perform type inference on an expression.
 #[cfg(test)]
@@ -408,7 +409,7 @@ fn typed_expr(code: &str) -> TypedExpr {
 /// Check whether a function has a vectorized Merge call. We'll use this to check whether function
 /// bodies got vectorized.
 #[cfg(test)]
-pub fn has_vectorized_merge(expr: &TypedExpr) -> bool {
+fn has_vectorized_merge(expr: &TypedExpr) -> bool {
     let mut found = false;
     expr.traverse(&mut |ref e| if let Merge { ref value, .. } = e.kind {
                            found |= value.ty.is_simd();
